+++ front-end/src/pages/CreateAccount/CreateAccount.css
@@ -1,3 +1,13 @@
+header {
+    background-color: #D9D9D9;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    color: black;
+    margin-bottom: 5px;
+  }
+
 h1{
     text-align: center;
     font-size: calc(15px + 3vmin);
@@ -5,7 +15,8 @@ h1{
 
 *{
     box-sizing: border-box;
-    font-size: calc(8px + 2vmin);
+    font-size: calc(5px + 2vmin);
+    font-weight: 400;
 }
 
 html, body{
@@ -30,18 +41,40 @@ img{
 }
 
 .EmailAddress{
-    margin: calc(20px + 3vmin);
+    margin-top: calc(20px + 3vmin);
+    margin-bottom: calc(10px + 2vmin);
 }
 
-.ProfilePhoto{
-    margin: calc(20px + 3vmin);
+.UploadPP{
+    font-size: calc(6px + 1vmin);
+    margin: calc(10px + 1vmin);
+}
+
+.UploadButton{
+    font-size: calc(5px + 1vmin);
+    margin: calc(5px + 1vmin);
 }
 
 .Country, .State, .City, .Address{
     font-size: calc(2px + 2vmin);
-    margin: calc(8px + 1vmin);
+    margin: calc(5px + 1vmin);
+}
+
+.CountryInput, .StateInput, .CityInput, .AddressInput{
+    font-size: calc(1px + 2vmin);
+}
+
+.CreateButton{
+    font-size: calc(10px + 1vmin);
+    font-weight: 300;
+    margin: calc(10px + 1vmin);
 }
 
 .CreateAccount{
     text-align: center;
+}
+
+::placeholder{
+    color: #a9a9a9;
+    font-weight: 100;
 }
